split for form field and form set. Deletion of form.test.js
